% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_id.R
\name{check_id}
\alias{check_id}
\title{Check IDs}
\usage{
check_id(df, ..., max_depth = 3)
}
\arguments{
\item{df}{A data frame}

\item{...}{Columns within df. These can be quoted or NSE}

\item{max_depth}{The maximum number of columns to look for a unique key among.}
}
\description{
Test keys for uniqueness or try to find unique keys
}
\details{
The intent of this function is twofold. First, it will test if a set
of supplied columns are a unique key. By 'unique key' we mean that each
combination of those variables uniquely specifies a row in df.
Supply columns to \code{...} for this functionality. \cr \cr
The second purpose of check_id is to determine which sets of column(s)
if any, determine a unique key for \code{df}. If any do, they will be listed.
Othewise, the closest key(s) will be listed. The output in this case
is a data frame of statistics on the attempts.
}
\author{
Sven Halvorson
}
