% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{clear}
\alias{clear}
\title{Clear Workspace}
\usage{
clear(env = globalenv())
}
\arguments{
\item{env}{An environment}
}
\description{
Clears specified environment and console
}
\note{
This does not remove loaded package. Restart your R session before getting the final output.
}
\examples{
clear()
}
\author{
Sven Halvorson
}
