% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_multi.R
\name{coalesce_multi}
\alias{coalesce_multi}
\title{Coalesce Multiple Columns}
\usage{
coalesce_multi(df, pattern = stringr::fixed("."))
}
\arguments{
\item{df}{a data frame}

\item{pattern}{pattern to split column names along using \code{stringr::str_split_fixed}}
}
\value{
a data frame with coalesced columns
}
\description{
Coalesce columns matching the LHS when splitting by \code{pattern}. Columns
are coalesced from left to right as they appear in \code{df}
}
\note{
Columns that do not contain \code{pattern} but match another column after splitting
will STILL be coalesced. In the example, the columns \code{c(value, value.x, value.y)} are
coalesced when \code{(pattern = stringr::fixed('.')}.
}
\examples{
set.seed(1)
make_data = function(values){
  tibble::tibble(
    id = 1:10,
    value = sample(c(values, NA), size = 10, replace = TRUE),
    value2 = sample(c(toupper(values), NA), size = 10, replace = TRUE)
  )
}

# Notice the last record has missing values for both value.x and value.y
sample_data = list(
  make_data(letters[1]),
  make_data(letters[2]),
  dplyr::select(make_data(letters[3]), -value2)
) |>
  purrr::reduce(dplyr::full_join, by = 'id')
coalesce_multi(sample_data)
}
