% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_cases.R
\name{combine_cases}
\alias{combine_cases}
\title{Combine cases}
\usage{
combine_cases(.data, ..., .f)
}
\arguments{
\item{.data}{A data frame}

\item{...}{Unquoted grouping variables within \code{.data}}

\item{.f}{A summary function. Defaults to first non-missing value}
}
\value{
a tbl
}
\description{
Groups \code{.data} by \code{...} and applies \code{.f} to all other columns
}
\examples{
\dontrun{
example_data = tibble(
  id = sample(letters, size = 100, replace = TRUE),
  bin = sample(0:1, size = 100, replace = TRUE),
  normal = rnorm(n = 100)
) \%>\%
  arrange(id, bin, normal)

example_data[2:ncol(example_data)] = lapply(
  example_data[2:ncol(example_data)],
  function(x){x[rnorm(length(x)) < -0.5] = NA;x}
)
combine_cases(
  .data = example_data,
  id
)
}
}
\author{
Sven Halvorson
}
