% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{crosswalk}
\alias{crosswalk}
\title{Create Crosswalk}
\usage{
crosswalk(df, ..., all_combos = FALSE)
}
\arguments{
\item{df}{A data frame}

\item{...}{Coluumns within \code{df}}

\item{all_combos}{Should combinations of \code{...} not present in \code{df} be created?}
}
\value{
Code to a rstudio script that generates the crosswalk.
}
\description{
Manually create a crosswalk from one or more columns of a data frame.
}
\details{
Often text data needs to be binned or converted. Doing this via
large sets of regular expressions sometimes works but often it's faster just
to make a crosswalk yourself. \code{crosswalk} creates the unique combination
of one or more columns within a data frame, allows the user to edit the 'new
values' and returns code to produce the crosswalk.

You must highlight a single line within an R studio script for this to execute. This
is intended to help you not delete your code.
}
\author{
Sven Halvorson
}
