% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtypes.R
\name{dtypes}
\alias{dtypes}
\title{Data types}
\usage{
dtypes(df)
}
\arguments{
\item{df}{A \code{data.frame}}
}
\value{
A tibble of the types and classes of the columns of \code{df}
}
\description{
Data types
}
\note{
I wanted a version of the pandas function so here it is
}
\examples{
dtypes(iris)
}
\author{
Sven Halvorson
}
