% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupes.R
\name{dupes}
\alias{dupes}
\title{Two-way Duplicates}
\usage{
dupes(x)
}
\arguments{
\item{x}{a vector, data frame, array, or \code{NULL}}
}
\value{
A logical object of the same dimensions as \code{x}
}
\description{
Two-way Duplicates
}
\examples{
x = c(1:4, 1, 2, NA)
dupes(x)
}
\author{
Sven Halvorson
}
