% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cols.R
\name{find_cols}
\alias{find_cols}
\title{Find Columns}
\usage{
find_cols(df, ..., value = TRUE, ignore.case = TRUE)
}
\arguments{
\item{df}{A data frame}

\item{...}{Regular expressions, quoted or not}

\item{value}{Should the column names or positions be returned?}

\item{ignore.case}{Should case be ignored in regular expression matching?}
}
\description{
Find columns in a data frame by regular expression
}
\note{
This is mostly for interactive use
}
\examples{
find_cols(mtcars, ar, pg)
find_cols(mtcars, 'ar', value = FALSE)
}
\author{
Sven Halvorson
}
