% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_checkbox.R
\name{format_checkbox}
\alias{format_checkbox}
\title{Format Checkbox Columns}
\usage{
format_checkbox(df, relabel = TRUE)
}
\arguments{
\item{df}{A data frame}

\item{relabel}{Should the value label be set as the variable label?}
}
\value{
A data frame
}
\description{
Modify a data frame representing a 'select all that apply'
style survey question. Makes implicit negative responses explicit while
keeping non-responses as is.
}
\details{
The steps taken are:
\enumerate{
  \item Set the variable label as the (presumably singular) value label if \code{relabel = TRUE}
  \item Find the observations that are missing across all columns
  \item Remove value labels, replace missing values with zeroes
  \item Replace values with NA for completely missing observations
}
}
\examples{
set.seed(0)
checkbox_example = tibble(
  TREATMENT = rep(c('treat', 'control'), each = 20),
  Q1 = sample(c(1, NA), size = 40, replace = TRUE),
  Q2 = sample(c(1, NA), size = 40, replace = TRUE),
  Q3 = sample(c(1, NA), size = 40, replace = TRUE)
)
# Note first & third rows are all missing

# Set some value labels:
for(i in 1:3){
  lab = 1
  names(lab) = LETTERS[i]
  expss::val_lab(checkbox_example[paste0('Q',i)]) = lab
}

# Make a new dataframe of just the formatted checkbox questions:
formatted_checkbox = checkbox_example \%>\%
  select(all_of(paste0('Q', 1:3))) \%>\%
  format_checkbox()

attributes(formatted_checkbox[['Q1']])

# Also can do assignment for a chunk of the data frame but it's a little trickier
checkbox_example[,paste0('Q', 1:3)] = format_checkbox(checkbox_example[,paste0('Q', 1:3)])
checkbox_example
}
\author{
Sven Halvorson (svenpubmail@gmail.com)
}
