% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ci.R
\name{format_ci}
\alias{format_ci}
\title{Format Confidence Intervals}
\usage{
format_ci(
  point,
  lower,
  upper,
  null_value = NA_real_,
  digits = NA_real_,
  max_its = 4,
  unacceptables = NA_real_
)
}
\arguments{
\item{point, lower, upper}{Point estimates, lower bounds, and upper bounds.}

\item{null_value}{A vector of null values to be compared to.}

\item{digits}{A vector of digits to begin rounding at.}

\item{max_its}{The maximum number of additiona digits to give}
}
\description{
Round and format confidence intervals for presentation.
}
\details{
The goal of this function is to round confidence intervals
for presentation. You can supply digits to round to or let the function guess.
It will try to improve the CI until it hits the \code{max_its}. The function
attempts to have the point estimate and each end of the CI as different numbers.
If you supply a \code{null_value} it will try not to use that as either endpoint
and keep the null value on the same side of the CI as the unformatted value is.
If \code{unacceptables} is supplied, these values will not be allowed as
endpoints for the CI.
}
\examples{
# Set up a dataset for a logistic regression model:
df = nycflights13::flights \%>\%
  mutate(on_time = dep_delay <= 0)

analysis_table = glm(
  formula = on_time ~ month + sched_dep_time + origin + air_time,
  data = df,
  family = binomial
) \%>\%
  broom::tidy(
    conf.int = TRUE
  )

cis = format_ci(
  point = analysis_table['estimate'],
  lower = analysis_table['conf.low'],
  upper = analysis_table['conf.high'],
  null_value = 1,
  unacceptables = 0
)

analysis_table \%>\%
  select(
    term,
    estimate,
    conf.low,
    conf.high
  ) \%>\%
  bind_cols(cis)

}
\author{
Sven Halvorson
}
