% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_file.R
\name{get_latest_file}
\alias{get_latest_file}
\title{Get latest file}
\usage{
get_latest_file(path, pattern = NULL, full.names = TRUE, ...)
}
\arguments{
\item{path}{Path to a directory to search within}

\item{pattern}{Pattern of files to look for}

\item{full.names}{Should the complete paths be returned?}

\item{...}{Other arguments for \code{list.files}}
}
\value{

}
\description{
Get latest file
}
\details{
I have a habit of putting the YYYYMMDD date into the files that I save
in order to keep old versions. This function helps find the last version of a file
by extracting those 8 digit strings, finding a maximumn, and returning the file(s)
that have that max in their name. Note that this will totally crash and burn
if the dates are in another format or there are other 8+ digit sequences in the file names.
}
\examples{
get_latest_file(
  path = '//psi-svr-filer.uoregon.edu/P50/radx_latinx_collection/data/external/site_info',
  pattern = 'cleaned_schedule'
)
}
\seealso{
\code{list.files}
}
\author{
Sven Halvorson (svenpubmail@gmail.com)
}
