% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggslideshow.R
\name{ggslideshow}
\alias{ggslideshow}
\title{Slideshow of ggplots}
\usage{
ggslideshow(plots, delay = 3)
}
\arguments{
\item{plots}{a list of ggplot objects}

\item{delay}{the delay in seconds}
}
\description{
Slideshow of ggplots
}
\note{
This is mostly for interactive use. It just prints the plots in order#'
}
\examples{
make_iris_plot = function(sp){
  iris \%>\%
    filter(Species == sp) \%>\%
    ggplot(aes(x = Sepal.Width, y = Sepal.Length)) +
    geom_point() +
    labs(title = sp)
}

unique(iris$Species) \%>\%
  lapply(make_iris_plot) \%>\%
  ggslideshow()
}
\author{
Sven Halvorson
}
