% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled_to_character.R
\name{labelled_to_character}
\alias{labelled_to_character}
\title{Convert labelled numerics to character}
\usage{
labelled_to_character(x)
}
\arguments{
\item{x}{a labelled numeric}
}
\value{
a character vector
}
\description{
Convert labelled numerics to character
}
\examples{
labelled_num = sample(1:3, size = 10, replace = TRUE)
expss::val_lab(labelled_num) = c(
  'Tyre' = 1,
  'Byblos' = 2,
  'Ugarit' = 3
)
table(
  labelled_num,
  labelled_to_character(labelled_num)
)

}
\author{
Sven Halvorson (svenpubmail@gmail.com)
}
