% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled_to_continuous.R
\name{labelled_to_continuous}
\alias{labelled_to_continuous}
\title{Labelled vectors to continuous}
\usage{
labelled_to_continuous(x)
}
\arguments{
\item{x}{an atomic numeric vector with the attribute "labels"}
}
\value{
a numeric vector
}
\description{
Convert a vector to the numeric version of its value labels
}
\details{
\code{labelled_to_continuous} first converts the values \code{x} to
their associated labels. Then it returns \code{as.numeric(x)}. The intended
use case of this is when a continuos measure was ingested as a categorical data type
and the labels contain the desired values (wrong recodes).
}
\examples{
labelled_vector = 1:4
expss::var_lab(labelled_vector) = 'How many times have you read Enuma Elish?'
expss::val_lab(labelled_vector) = c(
  'Prefer not to answer' = 1,
  '0' = 2,
  '1' = 3,
  '2' = 4
)
labelled_to_continuous(labelled_vector)

}
\author{
Sven Halvorson (svenpubmail@gmail.com)
}
