% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled_to_factor.R
\name{labelled_to_factor}
\alias{labelled_to_factor}
\title{Convert labelled numerics to factors}
\usage{
labelled_to_factor(x)
}
\arguments{
\item{x}{an atomic numeric with the attribute "labels". Probably from \code{expss::val_lab}}
}
\value{
a factor
}
\description{
Convert labelled numerics to factors
}
\note{
This is a bit of a quirky solution because the integers underlying
a factor are default set to 1 through the number of unique items. This means
that without saving the original value labels
}
\examples{
city_labels = c(
  'Urruk' = 2,
  'Babylon' = 3,
  'Larsa' = 4,
  'Mari' = 5
)

set.seed(0)
cities = sample(2:5, size = 10, replace = TRUE)

expss::val_lab(cities) = city_labels

labelled_to_factor(cities)

# NOTE the difference:
as.numeric(labelled_to_factor(cities)) == cities
}
\author{
Sven Halvorson (svenpubmail@gmail.com)
}
