% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_periods.R
\name{merge_periods}
\alias{merge_periods}
\title{Merge time intervals}
\usage{
merge_periods(
  df,
  start_time,
  end_time,
  ...,
  tolerance = 0,
  units = "minutes",
  simplify = TRUE,
  plots = FALSE
)
}
\arguments{
\item{df}{A \code{dataframe}}

\item{start_time}{The unquoted name of the starting time column in \code{df}}

\item{end_time}{The unquoted name of the ending time column in \code{df}}

\item{...}{Optional grouping columns, unquoted}

\item{tolerance}{A number specifying how far apart start and end times can be
to combine.}

\item{units}{The units for tolerance. One of \code{c('seconds', 'minutes', 'hours', 'days', 'years')}}

\item{simplify}{Should the results be simplified and joined?}

\item{plots}{Should plots be generated for each group?}
}
\description{
Join adjacent or redundant time intervals across ids
}
\examples{
data(periods_data)
merged_periods = merge_periods(
  periods_data,
  start_time = ts_start,
  end_time = ts_end,
  id,
  tolerance = 3,
  units = 'minutes',
  simplify = FALSE,
  plots = TRUE
)

SvenR::ggslideshow(merged_periods$plot)
}
\author{
Sven Halvorson
}
