% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nin.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Not In}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{values to be matched}

\item{y}{values to be matched against}
}
\description{
Inverse of infix \code{\%in\%}
}
\examples{
'a' \%nin\% letters
'a' \%nin\% LETTERS
}
\seealso{
match
}
\author{
Somebody else, not sure who
}
