% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_mrn.R
\name{pad_mrn}
\alias{pad_mrn}
\title{Pad MRN}
\usage{
pad_mrn(mrn)
}
\arguments{
\item{mrn}{Vector of MRNs}
}
\value{
A character vector of MRNs each having 8 characters and with leading zeroes
}
\description{
Pad MRN
}
\note{
This shouldn't always be used as some local MRNs are not supposed to have 8 digits
}
\examples{
pad_mrn(c(12345678, 7654321))
}
\author{
Sven Halvorson
}
