% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretend_working.R
\name{pretend_working}
\alias{pretend_working}
\title{Pretend to Work}
\usage{
pretend_working(minutes = 5)
}
\arguments{
\item{minutes}{run time}
}
\description{
Creates poorly veiled wizbangs that might make it look
like someting important is happening on your computer
}
\author{
Sven Halvorson
}
