% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_iqr.R
\name{pretty_iqr}
\alias{pretty_iqr}
\title{Pretty IQR}
\usage{
pretty_iqr(x, digits = 1, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{digits}{Digits to round to}

\item{na.rm}{Remove NA values?}
}
\description{
Formatted character version of median, Q1, Q3
}
\examples{
pretty_iqr(mtcars$mpg)
}
\author{
Sven Halvorson
}
