% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rin.R
\name{\%rin\%}
\alias{\%rin\%}
\title{Reverse In}
\usage{
x \%rin\% y
}
\arguments{
\item{x}{values to be matched against}

\item{y}{values to be matched}
}
\description{
Reverse of infix \code{\%in\%}
}
\examples{
'a' \%rin\% letters
LETTERS \%rin\% 'a'
}
\seealso{
match
}
\author{
Sven Halvorse
}
