% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup_join.R
\name{rollup_join}
\alias{rollup_join}
\title{Rollup join}
\usage{
rollup_join(dfs, ..., how = "left")
}
\arguments{
\item{dfs}{A named list of \code{data.frames}}

\item{...}{The keys to join on, unquoted}

\item{how}{The merging style. One of \code{c('left', 'right', 'inner', 'full')}}
}
\description{
Join a list of data sets and check for preserving unique ids.
The idea for this function is that I frequently assemble many data frames
and wish to combine them at the end of the script for export. This helps
make sure you preserve the uniqueness of the ids and tells you where the
problem is occurring.
}
\note{
It's easy to get a named list if you use mget
}
\examples{

d1 = tibble(id = 1:10,
            a = sample(letters, 10, replace = TRUE))
d2 = tibble(id = 1:5,
            b = sample(letters, 5, replace = TRUE))
d3 = tibble(id = c(1:3, 1),
            c = sample(letters, 4, replace = TRUE))
rollup_join(mget(paste0('d', 1:2)), id)
rollup_join(mget(paste0('d', 1:3)), id)
}
\author{
Sven Halvorson
}
