% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_ci.R
\name{round_ci}
\alias{round_ci}
\title{Format CIs}
\usage{
round_ci(estimate, lower, upper, digits = 2, max_its = 4, null_value)
}
\arguments{
\item{estimate, }{lower, upper The point estimate, lower bound, and upper bound.}

\item{digits}{A vector of digits to begin rounding at.}

\item{max_its}{The maximum number of additiona digits to give}

\item{null_value}{A vector of null values to be compared to. If}
}
\description{
Round and format confidence intervals
}
\author{
Sven Halvorson
}
