% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_join.R
\name{safe_join}
\alias{safe_join}
\title{One-to-one left join}
\usage{
safe_join(x, y, ...)
}
\arguments{
\item{x, y}{tbls to join}

\item{...}{other arguments to be passed to \code{left_join}}
}
\value{
a tbl
}
\description{
Just a simple wrapper around \code{dplyr::left_join} to stop myself
from accidentally many-to-one joining when I don't expect to.
}
\note{
Saftey first!
}
\examples{
\dontrun{
df1 = tibble(
  x = 1:5,
  y = letters[1:5],
  z = LETTERS[1:5]
)

df2 = tibble(
  x = c(1:4, 4),
  j = LETTERS[5:1]
)

safe_join(df1, df2)
safe_join(df1, df2, by = c('z' = 'j'), suffix = c('_left', '_right'))
}
}
\author{
Sven Halvorson
}
