% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scols.R
\name{scols}
\alias{scols}
\title{Sort Columns}
\usage{
scols(df)
}
\arguments{
\item{df}{a data frame}
}
\description{
Useful for finding colum names
}
\note{
For interactive use
}
\examples{
scols(mtcars)
}
\author{
Sven Halvorson
}
