% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_list.R
\name{sql_list}
\alias{sql_list}
\title{SQL List}
\usage{
sql_list(x, quote = TRUE, wildcard = FALSE)
}
\arguments{
\item{x}{Atomic character or numeric vector}

\item{quote}{Should each entry be surrounded with single quotes?}

\item{wildcard}{Should each entry be surrounded with percent signs? Used for wildcard matching}
}
\description{
Format an atomic vector for use with a SQL 'IN' statements
}
\examples{
paste(
  'SELECT nonsense',
  'FROM your_table',
  'WERE nonsense LIKE ANY',
  sql_list(letters[1:5], wildcard = TRUE),
  ';'
)
}
\author{
Sven Halvorson
}
