% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_na.R
\name{sum_na}
\alias{sum_na}
\title{Sum Missing Values}
\usage{
sum_na(x)
}
\arguments{
\item{x}{a vector, matrix, or data.frame}
}
\description{
Sum Missing Values
}
\examples{
sum_na(c(1, NA, 2, NA, 3, NA))
}
\author{
Sven Halvorson
}
