% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twa.R
\name{twa}
\alias{twa}
\title{Time Weighted Averages}
\usage{
twa(df, value_var, time_var, ..., method = "trapezoid", ref = 0,
  ref_dir = "raw")
}
\arguments{
\item{df}{a data frame, assumed to be in a 'long form' with time values in a single column}

\item{value_var}{the value to be avaraged, a column within df}

\item{time_var}{the time variable for weighting. Either \code{POSIXct} or numeric.}

\item{...}{grouping variables within df}

\item{method}{method to compute TWA, one of \code{c('trapezoid', 'left', 'right')}}

\item{ref}{a value to compute the TWA relative to}

\item{ref_dir}{the direction to compute the average relative to ref, one of \code{c('raw', 'above', 'below', 'about')}}
}
\value{
a data frame containing any grouping variables, the computed twa, and some other summary statistics
}
\description{
This function computes time weighted averages of \code{value_var} across zero or more groups.
Averages can be computed by trapezoids or left/right endpoints. Time weighted averages can be computed
with raw values or relative to a reference. Calculations can be above below, or about the supplied
reference value. By default, they are computed as raw.
}
\details{
If multiple rows have the same grouping variables and time, the median value will be selected.
Rows with missing. If only one value within a grouping level is supplied, that value will be returned.d
\code{twa} also computes the total time elapsed, the largest and smallest gap between consecutive points,
and the number of values received, used, and omitted. When \code{time_var} is \code{POSIXct}, \code{twa}
will default to minutes.
}
\examples{
start_date = ymd_hms('2019-01-01 00:00:00')
time_dat = tibble(id = c(1, 1, 1, 1, 2, 2),
                  val = c(4, 6, 8, 6, 1, NA),
                  t = minutes(c(0, 10, 15, 45, 0, 10)) + start_date,
                  t2 = 1:6)
twa(df = time_dat, value_var = val, time_var = t, id)
twa(df = time_dat, value_var = val, time_var = t2, method = 'left', ref = 7, ref_dir = 'below')
}
\author{
Sven Halvorson
}
