% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_view.R
\name{variable_view}
\alias{variable_view}
\title{SPSS Style Variable Viewer}
\usage{
variable_view(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
a summary table of \code{df}
}
\description{
Create a table of column names and attributes.
}
\examples{
example_df = tibble::tibble(
  Goddess = c(1, 2, 3),
  Civilization = c(
    'Babylonian',
    'Sumerian',
    'Sumerian'
  ),
  Importance = runif(3),
  Date = lubridate::mdy(
    c(
      '01-30-1988',
      '05-22-2006',
      '11-02-2017'
    )
  )
)

expss::val_lab(example_df[['Goddess']]) = c(
  'Tiamat' = 1,
  'Inanna' = 2,
  'Ninhursag' = 3
)

expss::var_lab(example_df[['Date']]) = 'A random date'

example_df \%>\%
  variable_view() \%>\%
  View()
}
\author{
Sven Halvorson (svenpubmail@gmail.com)
}
