% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ventile.R
\name{ventile}
\alias{ventile}
\title{Ventile (5-tiles)}
\usage{
ventile(x, quiet = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{quiet}{should it print observation count and missing rate?}
}
\value{
a numeric vector of 19 quantiles
}
\description{
Compute 5th, 10th, 15th ... percentiles
}
\examples{
ventile(
  x = sample(
    x = c(1:100, NA_real_),
    replace = TRUE,
    size = 1234,
    prob = 1:101
  ),
  quiet = FALSE
)
}
\author{
Sven Halvorson
}
