% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_ids.R
\name{count_ids}
\alias{count_ids}
\title{Count IDs}
\usage{
count_ids(df, ...)
}
\arguments{
\item{df}{A data frame}

\item{...}{Unquoted column names in \code{df}}
}
\value{
\code{df}, intended for side effect use.
}
\description{
Computes \code{n_distinct} on specified columns and gives row count.
}
\note{
Defaults to op_sys, visit_sys, and patient_sys if \code{...} is not supplied. Ungroups data frames with a warning.
}
\examples{
df = tibble(op_sys = 1:5,
            visit_sys = c(1, 1, 2, 3, 4),
            patient_sys = c(1, 1, 2, 2, 3))
count_ids(df)
count_ids(df, op_sys)
}
\author{
Sven Halvorson
}
