% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{crosswalk}
\alias{crosswalk}
\title{Create Crosswalk}
\usage{
crosswalk(df, ..., all_combos = FALSE)
}
\arguments{
\item{df}{A data frame}

\item{...}{Coluumns within \code{df}}

\item{all_combos}{Should combinations of \code{...} not present in \code{df} be created?}
}
\value{
Code to rstudio a script that generates the crosswalk.
}
\description{
Manually create a crosswalk from one or more columns of a data frame.
}
\examples{
iris \%>\%
  mutate(species_chr = Species) \%>\%
  crosswalk(species_chr)
}
\author{
Sven Halvorson
}
