% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cols.R
\name{find_cols}
\alias{find_cols}
\title{Find Columns}
\usage{
find_cols(df, pattern, value = TRUE, ignore.case = TRUE)
}
\arguments{
\item{df}{A data frame}

\item{pattern}{A regular expression}

\item{value}{Should the column names or positions be returned?}

\item{ignore.case}{Should case be ignored in regular expression matching?}
}
\description{
Find columns in a data frame by regular expression
}
\note{
This is mostly for interactive use
}
\author{
Sven Halvorson
}
