% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nin.R, R/rin.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Not In}
\usage{
x \%nin\% y

x \%nin\% y
}
\arguments{
\item{x}{values to be matched against}

\item{y}{values to be matched}
}
\description{
Inverse of infix \code{\%in\%}

Reverse of infix \code{\%in\%}
}
\examples{
'a' \%nin\% letters
'a' \%nin\% LETTERS
'a' \%nin\% letters
'a' \%nin\% LETTERS
}
\seealso{
match

match
}
\author{
Somebody else, not sure who

Sven Halvorse
}
