% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_mrn.R
\name{pad_mrn}
\alias{pad_mrn}
\title{Pad MRNs}
\usage{
pad_mrn(x)
}
\arguments{
\item{x}{atomic vector}
}
\description{
adds leading zeroes to numeric vector
}
\examples{
mrns = c(11111111, 222222, 33333)
pad_mrn(mrns)
}
\author{
Sven Halvorson
}
