% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_list.R
\name{sql_list}
\alias{sql_list}
\title{SQL List}
\usage{
sql_list(x, quote = TRUE)
}
\arguments{
\item{x}{Atomic character or numeric vector}
}
\description{
Format an atomic vector for use with a SQL 'IN' statements
}
\examples{
paste(
  'SELECT nonsense',
  'FROM your_table',
  'WERE nonsense IN',
  sql_list(letters[1:5]),
  ';'
)
}
\author{
Sven Halvorson
}
